mods/dotfiles/global_python_scripts/pyproject.toml
[project]
name = "global-python-scripts"
version = "0.1.0"
description = "a repository of globally installed python scripts"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "aider-chat",
  "PyYAML"
]
